*** empty log message ***

+++ zerofpsv2/src/dark_metropolis/dark_metropolis_main.cpp
@@ -372,7 +372,7 @@ void DarkMetropolis::MoveCamera(Vector3 pos)
 void DarkMetropolis::Input()
 {
 	//get mouse
-	int x,z;		
+	float x,z;		
 	m_pkInputHandle->RelMouseXY(x,z);	
 
 	// Byt tillbaks till bakgrundsmusik om det har gått 10 sek efter det
