+++ zerofpsv2/src/fps_test/dark_metropolis_gui.cpp
@@ -92,7 +92,7 @@ void DarkMetropolis::GUI_OnIdle()
 
 			Rect rc = m_pkGamePlayInfoLabel->GetScreenRect();
 
-			int mx,my;
+			float mx,my;
 			m_pkInputHandle->MouseXY(mx,my);
 
 			my+=rc.Height()+16;
